add  shared vote rules  deployed
